Add StateRef variant for state helpers
There may be cases where you want a persistent or stored state but also have a ref to the value (to update outside of the react loop, for instance). This commit adds those variants so you don’t have to do it manually. This also fixes a lint error and updates the deploy script to work. This also adds a fix script for ease of use.