[0.1.4] Go back to tuple type of types for Merge/Union
The UnionToType helpers is a little unpredictable (it's not really a supported feature of typescript). This defaults to a tuple type of types, which is very consistent. If you need to pass it from a union type, we are exposing UnionToTuple for you for that reason. This also fixes an issue where undefined was being included in parsing the variadic tuple. This also runs the formatter again.