Fix up state/ref helpers parameters/issues
The state/ref helpers were a little wonky. This fixes them so they are consistent. This also allows the initial values to be set with a function (like the react’s `useState`) and makes the setters a `SetStateAction` so that we can utilize the current value without adding it as a dependency.