[2.1.2] Fix DST issue in `parseDate`
The `parseDate` helper was using the current local timezone offset, but the interpreters will use the timezone offset of the date in question. That caused a mismatch where if the timezone offset of the date being parsed was different than the local time zone offset (like during DST, etc), the dates wouldn't match.