[2.1.3] Fix `parseDate` for pre-standardization dates
Dates that were before November 18th, 1883 used a wild-wild-west amalgamation of standards. That's when the standardized time zones we know today [was adopted](https://time.com/5730146/standardized-time/). This fixes an issue our parser had with those early dates. The side affect is that this also fixes a bug with parsing dates from the years 0-99.