[3.6.0] Add manuel store option to memoize
Users may want to store the memoize cache elsewhere than the local memory, perhaps like in a shared memory or even a db. This will allow you to do that by letting you manually define the cache via a set of functions.

We also add tests for memoize and other utilized helpers.