[4.3.0] Add `sort` helper
This helper allows you to easily sort using a specific value without calculating that value many times in a sorter function. The built-in sort will call the function a bunch of times, so if the sort property is expensive, this helper will allow you to calculate each value just once and then sort on it.